Move everything to 64 bits, add day 9 part 1

+++ src/day1.rs
@@ -1,28 +1,28 @@
 use crate::util;
 
-pub fn part1() -> i32 {
+pub fn part1() -> i64 {
     util::lines_from_path("data/d1.txt")
         .map(|x| match x {
-            Ok(line) => mass_to_fuel(line.parse::<i32>().unwrap()),
+            Ok(line) => mass_to_fuel(line.parse::<i64>().unwrap()),
             _ => 0,
         })
-        .sum::<i32>()
+        .sum::<i64>()
 }
 
-pub fn part2() -> i32 {
+pub fn part2() -> i64 {
     util::lines_from_path("data/d1.txt")
         .map(|x| match x {
-            Ok(line) => mass_to_fuel_inclusive(line.parse::<i32>().unwrap()),
+            Ok(line) => mass_to_fuel_inclusive(line.parse::<i64>().unwrap()),
             _ => 0,
         })
-        .sum::<i32>()
+        .sum::<i64>()
 }
 
-pub fn mass_to_fuel(mass: i32) -> i32 {
+pub fn mass_to_fuel(mass: i64) -> i64 {
     (mass / 3) - 2
 }
 
-pub fn mass_to_fuel_inclusive(mass: i32) -> i32 {
+pub fn mass_to_fuel_inclusive(mass: i64) -> i64 {
     let mut fuel = mass_to_fuel(mass);
 
     let mut added_fuel = mass_to_fuel(fuel);
